Revert "Fetch fresh snapshot from peer when joining (#6700)"
This reverts commit 44a669d2091d019f409661c1ea7e0c7f00c5a03d.
