+++ doc/host_config_schema/cchost_config.json
@@ -396,11 +396,6 @@
                     "type": "boolean",
                     "default": true,
                     "description": "Whether to follow redirects to the primary node of the existing service to join"
-                  },
-                  "fetch_recent_snapshot": {
-                    "type": "boolean",
-                    "default": true,
-                    "description": "Whether to ask the target for a newer snapshot before joining. The node will ask the target what their latest snapshot is, and if that is later than what the node has locally, will fetch it via RPC before launching. Should generally only be turned off for specific test cases"
                   }
                 },
                 "required": ["target_rpc_address"],
