+++ src/host/main.cpp
@@ -24,8 +24,7 @@
 #include "process_launcher.h"
 #include "rpc_connections.h"
 #include "sig_term.h"
-#include "snapshots/fetch.h"
-#include "snapshots/snapshot_manager.h"
+#include "snapshots.h"
 #include "ticker.h"
 #include "time_updater.h"
 
@@ -377,7 +376,7 @@ int main(int argc, char** argv)
       config.ledger.read_only_directories);
     ledger.register_message_handlers(bp.get_dispatcher());
 
-    snapshots::SnapshotManager snapshots(
+    asynchost::SnapshotManager snapshots(
       config.snapshots.directory,
       writer_factory,
       config.snapshots.read_only_directory);
@@ -507,6 +506,8 @@ int main(int argc, char** argv)
     enclave_config.writer_config = writer_config;
 
     ccf::StartupConfig startup_config(config);
+
+    startup_config.snapshot_tx_interval = config.snapshots.tx_count;
 
     if (startup_config.attestation.snp_security_policy_file.has_value())
     {
@@ -689,62 +690,22 @@ int main(int argc, char** argv)
       config.command.type == StartType::Join ||
       config.command.type == StartType::Recover)
     {
-      auto latest_local_snapshot = snapshots.find_latest_committed_snapshot();
-
-      if (
-        config.command.type == StartType::Join &&
-        config.command.join.fetch_recent_snapshot)
+      auto latest_committed_snapshot =
+        snapshots.find_latest_committed_snapshot();
+      if (latest_committed_snapshot.has_value())
       {
-        // Try to fetch a recent snapshot from peer
-        const size_t latest_local_idx = latest_local_snapshot.has_value() ?
-          snapshots::get_snapshot_idx_from_file_name(
-            latest_local_snapshot->second) :
-          0;
-        auto latest_peer_snapshot = snapshots::fetch_from_peer(
-          config.command.join.target_rpc_address,
-          config.command.service_certificate_file,
-          latest_local_idx);
+        auto& [snapshot_dir, snapshot_file] = latest_committed_snapshot.value();
+        startup_snapshot = files::slurp(snapshot_dir / snapshot_file);
 
-        if (latest_peer_snapshot.has_value())
-        {
-          LOG_INFO_FMT(
-            "Received snapshot {} from peer (size: {}) - writing this to disk "
-            "and using for join startup",
-            latest_peer_snapshot->snapshot_name,
-            latest_peer_snapshot->snapshot_data.size());
-
-          const auto dst_path = fs::path(config.snapshots.directory) /
-            fs::path(latest_peer_snapshot->snapshot_name);
-          if (files::exists(dst_path))
-          {
-            LOG_FATAL_FMT(
-              "Unable to write peer snapshot - already have a file at {}. "
-              "Exiting.",
-              dst_path);
-            return static_cast<int>(CLI::ExitCodes::FileError);
-          }
-          files::dump(latest_peer_snapshot->snapshot_data, dst_path);
-          startup_snapshot = latest_peer_snapshot->snapshot_data;
-        }
+        LOG_INFO_FMT(
+          "Found latest snapshot file: {} (size: {})",
+          snapshot_dir / snapshot_file,
+          startup_snapshot.size());
       }
-
-      if (startup_snapshot.empty())
+      else
       {
-        if (latest_local_snapshot.has_value())
-        {
-          auto& [snapshot_dir, snapshot_file] = latest_local_snapshot.value();
-          startup_snapshot = files::slurp(snapshot_dir / snapshot_file);
-
-          LOG_INFO_FMT(
-            "Found latest local snapshot file: {} (size: {})",
-            snapshot_dir / snapshot_file,
-            startup_snapshot.size());
-        }
-        else
-        {
-          LOG_INFO_FMT(
-            "No snapshot found: Node will replay all historical transactions");
-        }
+        LOG_INFO_FMT(
+          "No snapshot found: Node will replay all historical transactions");
       }
     }
