+++ src/node/node_state.h
@@ -2647,7 +2647,7 @@ namespace ccf
         throw std::logic_error("Snapshotter already initialised");
       }
       snapshotter = std::make_shared<Snapshotter>(
-        writer_factory, network.tables, config.snapshots.tx_count);
+        writer_factory, network.tables, config.snapshot_tx_interval);
     }
 
     void read_ledger_entries(::consensus::Index from, ::consensus::Index to)
