+++ tests/config.jinja
@@ -50,8 +50,7 @@
     {
       "retry_timeout": "{{ join_timer }}",
       "target_rpc_address": "{{ target_rpc_address }}",
-      "follow_redirect": {{ follow_redirect|tojson }},
-      "fetch_recent_snapshot": {{ fetch_recent_snapshot|tojson }}
+      "follow_redirect": {{ follow_redirect|tojson }}
     },
     "recover": {
       "initial_service_certificate_validity_days": {{ initial_service_cert_validity_days }},
