+++ tests/e2e_operations.py
@@ -209,105 +209,6 @@ def test_large_snapshot(network, args):
     )
 
 
-def test_snapshot_access(network, args):
-    primary, _ = network.find_primary()
-
-    snapshots_dir = network.get_committed_snapshots(primary)
-    snapshot_name = ccf.ledger.latest_snapshot(snapshots_dir)
-    snapshot_index, _ = ccf.ledger.snapshot_index_from_filename(snapshot_name)
-
-    with open(os.path.join(snapshots_dir, snapshot_name), "rb") as f:
-        snapshot_data = f.read()
-
-    with primary.client() as c:
-        r = c.head("/node/snapshot", allow_redirects=False)
-        assert r.status_code == http.HTTPStatus.PERMANENT_REDIRECT.value, r
-        assert "location" in r.headers, r.headers
-        location = r.headers["location"]
-        assert location == f"/node/snapshot/{snapshot_name}"
-        LOG.warning(r.headers)
-
-        for since, expected in (
-            (0, location),
-            (1, location),
-            (snapshot_index // 2, location),
-            (snapshot_index - 1, location),
-            (snapshot_index, None),
-            (snapshot_index + 1, None),
-        ):
-            for method in ("GET", "HEAD"):
-                r = c.call(
-                    f"/node/snapshot?since={since}",
-                    allow_redirects=False,
-                    http_verb=method,
-                )
-                if expected is None:
-                    assert r.status_code == http.HTTPStatus.NOT_FOUND, r
-                else:
-                    assert r.status_code == http.HTTPStatus.PERMANENT_REDIRECT.value, r
-                    assert "location" in r.headers, r.headers
-                    actual = r.headers["location"]
-                    assert actual == expected
-
-        r = c.head(location)
-        assert r.status_code == http.HTTPStatus.OK.value, r
-        assert r.headers["accept-ranges"] == "bytes", r.headers
-        total_size = int(r.headers["content-length"])
-
-        a = total_size // 3
-        b = a * 2
-        for start, end in [
-            (0, None),
-            (0, total_size),
-            (0, a),
-            (a, a),
-            (a, b),
-            (b, b),
-            (b, total_size),
-            (b, None),
-        ]:
-            range_header_value = f"{start}-{'' if end is None else end}"
-            r = c.get(location, headers={"range": f"bytes={range_header_value}"})
-            assert r.status_code == http.HTTPStatus.PARTIAL_CONTENT.value, r
-
-            expected = snapshot_data[start:end]
-            actual = r.body.data()
-            assert (
-                expected == actual
-            ), f"Binary mismatch, {len(expected)} vs {len(actual)}:\n{expected}\nvs\n{actual}"
-
-        for negative_offset in [
-            1,
-            a,
-            b,
-        ]:
-            range_header_value = f"-{negative_offset}"
-            r = c.get(location, headers={"range": f"bytes={range_header_value}"})
-            assert r.status_code == http.HTTPStatus.PARTIAL_CONTENT.value, r
-
-            expected = snapshot_data[-negative_offset:]
-            actual = r.body.data()
-            assert (
-                expected == actual
-            ), f"Binary mismatch, {len(expected)} vs {len(actual)}:\n{expected}\nvs\n{actual}"
-
-        # Check error handling for invalid ranges
-        for invalid_range, err_msg in [
-            (f"{a}-foo", "Unable to parse end of range value foo"),
-            ("foo-foo", "Unable to parse start of range value foo"),
-            (f"foo-{b}", "Unable to parse start of range value foo"),
-            (f"{b}-{a}", "out of order"),
-            (f"0-{total_size + 1}", "larger than total file size"),
-            ("-1-5", "Invalid format"),
-            ("-", "Invalid range"),
-            ("-foo", "Unable to parse end of range offset value foo"),
-            ("", "Invalid format"),
-        ]:
-            r = c.get(location, headers={"range": f"bytes={invalid_range}"})
-            assert r.status_code == http.HTTPStatus.BAD_REQUEST.value, r
-            assert err_msg in r.body.json()["error"]["message"], r
-
-
 def split_all_ledger_files_in_dir(input_dir, output_dir):
     # A ledger file can only be split at a seqno that contains a signature
     # (so that all files end on a signature that verifies their integrity).
@@ -395,7 +296,6 @@ def run_file_operations(args):
                 test_forced_ledger_chunk(network, args)
                 test_forced_snapshot(network, args)
                 test_large_snapshot(network, args)
-                test_snapshot_access(network, args)
 
                 primary, _ = network.find_primary()
                 # Scoped transactions are not handled by historical range queries
