+++ tests/infra/clients.py
@@ -162,13 +162,7 @@ class Request:
         if self.headers:
             string += f" <blue>{truncate(str(self.headers), max_len=25)}</>"
         if self.body is not None:
-            if (
-                "content-type" in self.headers
-                and self.headers["content-type"] == "application/octet-stream"
-            ):
-                string += f"<binary: {len(self.body)} bytes>"
-            else:
-                string += escape_loguru_tags(f' {truncate(f"{self.body}")}')
+            string += escape_loguru_tags(f' {truncate(f"{self.body}")}')
 
         return string
 
@@ -273,15 +267,7 @@ class Response:
         status_color = (
             "red" if status_category in (4, 5) else "yellow" if redirect else "green"
         )
-
-        if (
-            "content-type" in self.headers
-            and self.headers["content-type"] == "application/octet-stream"
-        ):
-            body_s = f"<binary: {len(self.body)} bytes>"
-        else:
-            body_s = escape_loguru_tags(truncate(str(self.body)))
-
+        body_s = escape_loguru_tags(truncate(str(self.body)))
         # Body can't end with a \, or it will escape the loguru closing tag
         if len(body_s) > 0 and body_s[-1] == "\\":
             body_s += " "
