+++ tests/infra/remote.py
@@ -336,7 +336,6 @@ class CCFRemote(object):
         ignore_first_sigterm=False,
         node_container_image=None,
         follow_redirect=True,
-        fetch_recent_snapshot=True,
         max_uncommitted_tx_count=0,
         snp_security_policy_file=None,
         snp_uvm_endorsements_file=None,
@@ -534,7 +533,6 @@ class CCFRemote(object):
                 ignore_first_sigterm=ignore_first_sigterm,
                 node_address=remote_class.get_node_address(node_address),
                 follow_redirect=follow_redirect,
-                fetch_recent_snapshot=fetch_recent_snapshot,
                 max_uncommitted_tx_count=max_uncommitted_tx_count,
                 snp_security_policy_file=snp_security_policy_file,
                 snp_uvm_endorsements_file=snp_uvm_endorsements_file,
