+++ tests/reconfiguration.py
@@ -124,13 +124,7 @@ def test_add_node(network, args, from_snapshot=True):
             }
         )
     )
-    network.join_node(
-        new_node,
-        args.package,
-        args,
-        from_snapshot=from_snapshot,
-        fetch_recent_snapshot=from_snapshot,
-    )
+    network.join_node(new_node, args.package, args, from_snapshot=from_snapshot)
 
     # Verify self-signed node certificate validity period
     new_node.verify_certificate_validity_period(interface_name=operator_rpc_interface)
@@ -144,10 +138,9 @@ def test_add_node(network, args, from_snapshot=True):
     if not from_snapshot:
         with new_node.client() as c:
             s = c.get("/node/state")
-            body = s.body.json()
-            assert body["node_id"] == new_node.node_id
+            assert s.body.json()["node_id"] == new_node.node_id
             assert (
-                body["startup_seqno"] == 0
+                s.body.json()["startup_seqno"] == 0
             ), "Node started without snapshot but reports startup seqno != 0"
 
     # Now that the node is trusted, verify endorsed certificate validity period
@@ -875,7 +868,6 @@ def run_join_old_snapshot(args):
                     args.package,
                     args,
                     from_snapshot=True,
-                    fetch_recent_snapshot=False,
                     snapshots_dir=tmp_dir,
                     timeout=3,
                 )
@@ -899,7 +891,6 @@ def run_join_old_snapshot(args):
                     args.package,
                     args,
                     from_snapshot=False,
-                    fetch_recent_snapshot=False,
                     timeout=3,
                 )
             except infra.network.StartupSeqnoIsOld as e:
